+++ packages/xrpl/HISTORY.md
@@ -7,6 +7,9 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 ### Fixed
 * Fix `AccountRoot` ledger object to correctly parse `FirstNFTokenSequence` field.
 
+### Added
+* Implement support for XLS-85d Token Escrow amendment.
+
 ## 4.3.0 (2025-6-09)
 
 ### Added